requirements.txt
azure-identity>=1.25.2
azure-monitor-opentelemetry>=1.8.6
azure-search-documents>=11.6.0
azure-storage-blob>=12.28.0
fastapi>=0.129.0
firecrawl-py>=4.14.1
langchain>=1.2.10
langchain-aws>=0.2.15
langchain-community>=0.3.18
langchain-openai>=1.1.9
langgraph>=1.0.8
langsmith>=0.7.3
opentelemetry-instrumentation-fastapi>=0.60b0
psycopg2-binary>=2.9.11
pydantic>=2.12.5
pypdf>=6.7.0
python-dotenv>=1.2.1
redis>=7.1.1
requests>=2.32.5
sqlalchemy>=2.0.46
streamlit>=1.54.0
uvicorn>=0.40.0
yt-dlp>=2026.2.4
opensearch-py>=2.8.0
boto3>=1.36.0
opencv-python>=4.8.0
openai>=1.0.0
groq>=0.11.0
langchain-groq>=0.1.0
